LayoutTests/http/tests/websocket/tests/hybi/inspector/before-load.html

<!doctype html>
<html>
<head>
<script src="../../../../inspector/resources/inspector-test.js"></script>
<script>
function load()
{
    testRunner.waitUntilDone();

    const url = "ws://127.0.0.1:8880/websocket/tests/hybi/inspector/echo";
    window.testWebSocket = new WebSocket(url);
    window.testWebSocket.addEventListener("open", (event) => {
        // Only run the tests once the websocket has connected.
        runTest();
    });
}

function test()
{
    let suite = InspectorTest.createAsyncSuite("WebSocket.BeforeLoad");

    function logReadyState(readyState) {
        let readyStateString = "";

        switch (readyState) {
        case WI.WebSocketResource.ReadyState.Closed:
            readyStateString = "closed";
            break;

        case WI.WebSocketResource.ReadyState.Connecting:
            readyStateString = "connecting";
            break;

        case WI.WebSocketResource.ReadyState.Open:
            readyStateString = "open";
            break;

        }

        InspectorTest.log(`Resource has readyState "${readyStateString}"`);
    }

    suite.addTestCase({
        name: "WebSocket.BeforeLoad",
        description: "Test that the existing websocket is sent to the frontend when it is first opened.",
        test(resolve, reject) {
            const url = "ws://127.0.0.1:8880/websocket/tests/hybi/inspector/echo";
            let webSocketResources = WI.networkManager.mainFrame.resourceCollectionForType(WI.Resource.Type.WebSocket);
            InspectorTest.expectEqual(webSocketResources.size, 1, "Should have 1 WebSocketResource");

            let webSocketResource = Array.from(webSocketResources)[0];
            if (!webSocketResource) {
                reject("Missing WebSocket resource.");
                return;
            }

            InspectorTest.expectThat(webSocketResource instanceof WI.WebSocketResource, "Resource should be a WebSocket.");
            InspectorTest.expectEqual(webSocketResource.url, url, `Resource URL should be "${url}".`);
            InspectorTest.expectEqual(webSocketResource.statusCode, 101, "Resource status code should be 101.");
            InspectorTest.expectEqual(webSocketResource.statusText, "Switching Protocols", "Resource status text should be \"Switching Protocols\".");
            InspectorTest.expectThat(webSocketResource.requestHeaders.valueForCaseInsensitiveKey("Connection"), "Resource should have request header Connection.");
            InspectorTest.expectThat(webSocketResource.requestHeaders.valueForCaseInsensitiveKey("Sec-Websocket-Key"), "Resource should have request header Sec-Websocket-Key.");
            InspectorTest.expectThat(webSocketResource.requestHeaders.valueForCaseInsensitiveKey("Sec-Websocket-Version"), "Resource should have request header Sec-Websocket-Version.");
            InspectorTest.expectEqual(webSocketResource.requestHeaders.valueForCaseInsensitiveKey("Upgrade"), "websocket", "Resource should have request header Upgrade.");
            InspectorTest.expectThat(webSocketResource.responseHeaders.valueForCaseInsensitiveKey("Connection"), "Resource should have response header Connection.");
            InspectorTest.expectThat(webSocketResource.responseHeaders.valueForCaseInsensitiveKey("sec-websocket-accept"), "Resource should have response header sec-websocket-accept.");
            InspectorTest.expectEqual(webSocketResource.responseHeaders.valueForCaseInsensitiveKey("Upgrade"), "websocket", "Resource should have response header Upgrade.");
            InspectorTest.expectEqual(webSocketResource.readyState, WI.WebSocketResource.ReadyState.Open, "Resource should be open.");

            let expectedFrames = [
                {
                    data: "Hello World!",
                    isOutgoing: true,
                },
                {
                    data: "Hello World!",
                    isOutgoing: false,
                },
            ];

            let handleFrameAdded = (event) => {
                let frame = event.data;
                let expectedFrame = expectedFrames[webSocketResource.frames.length - 1];
                if (!expectedFrame)
                    return;

                InspectorTest.expectEqual(frame.data, expectedFrame.data, `Frame data should be '${expectedFrame.data}'`);
                InspectorTest.expectEqual(frame.isOutgoing, expectedFrame.isOutgoing, `Frame outgoing should be '${expectedFrame.isOutgoing}'`);

                if (webSocketResource.frames.length == expectedFrames.length) {
                    webSocketResource.removeEventListener(WI.WebSocketResource.Event.FrameAdded, handleFrameAdded);
                    resolve();
                }
            };
            webSocketResource.addEventListener(WI.WebSocketResource.Event.FrameAdded, handleFrameAdded);

            InspectorTest.evaluateInPage("window.testWebSocket.send(\"Hello World!\")");
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="load()">
<p>Tests that WebSockets created before the inspector loads are also tracked.</p>
</body>
</html>
